+++ src/mame/machine/esqpanel.cpp
@@ -19,369 +19,373 @@
 #include <thread>
 #include <list>
 
-class external_panel;
+namespace esqpanel {
 
-using external_panel_ptr = std::shared_ptr<external_panel>;
-typedef std::map<http_manager::websocket_connection_ptr, external_panel_ptr, std::owner_less<http_manager::websocket_connection_ptr>> connection_to_panel_map;
+	class external_panel;
 
-enum message_type {
-	UNKNOWN = 0,
-	ANALOG = 1 << 0,
-	BUTTON = 1 << 1,
-	CONTROL = 1 << 2,
-	DISPLAY = 1 << 3,
-	INFO = 1 << 4
-};
+	using external_panel_ptr = std::shared_ptr<external_panel>;
+	typedef std::map<http_manager::websocket_connection_ptr, external_panel_ptr, std::owner_less<http_manager::websocket_connection_ptr>> connection_to_panel_map;
 
-class external_panel : public std::enable_shared_from_this<external_panel>
-{
-public:
-	static int get_message_type(const char c)
-	{
-		switch(c)
-		{
-		case 'A':
-			return message_type::ANALOG;
-		case 'B':
-			return message_type::BUTTON;
-		case 'C':
-			return message_type::CONTROL;
-		case 'D':
-			return message_type::DISPLAY;
-		case 'I':
-			return message_type::INFO;
-		default:
-			return message_type::UNKNOWN;
-		}
-	}
-
-	external_panel() : m_send_message_types(0)
-	{
-		// printf("session: constructed\n");
-	}
-
-	int handle_control_message(const std::string &command)
-	{
-		int old = m_send_message_types;
-		std::istringstream is(command);
-		if (get_message_type(is.get()) != message_type::CONTROL)
-		{
-			return 0;
-		}
-
-		int n;
-		while (!is.eof()) {
-			char c = is.get();
-			int message_type = external_panel::get_message_type(c);
-			is >> n;
-			int send = (n != 0);
-			if (send)
-			{
-				m_send_message_types |= message_type;
-			}
-			else
-			{
-				m_send_message_types &= ~message_type;
-			}
-		}
-
-		return m_send_message_types ^ old;
-	}
-
-	int send_message_types()
-	{
-		return m_send_message_types;
-	}
-
-	bool send_display_data()
-	{
-		return m_send_message_types & message_type::DISPLAY;
-	}
-
-	bool send_analog_values()
-	{
-		return m_send_message_types & message_type::ANALOG;
-	}
-
-	bool send_buttons()
-	{
-		return m_send_message_types & message_type::BUTTON;
-	}
-
-private:
-	int m_send_message_types;
-};
-
-class esqpanel_external_panel_server
-{
-public:
-	enum websocket_opcode {
-		text = 1,
-		binary = 2
+	enum message_type {
+		UNKNOWN = 0,
+		ANALOG = 1 << 0,
+		BUTTON = 1 << 1,
+		CONTROL = 1 << 2,
+		DISPLAY = 1 << 3,
+		INFO = 1 << 4
 	};
-	esqpanel_external_panel_server(http_manager *webserver) :
-		m_server(webserver),
-		m_keyboard("unknown"),
-		m_version("1")
-	{
-		using namespace std::placeholders;
-		if (m_server->is_active()) {
-			m_server->add_endpoint("/esqpanel/socket",
-					std::bind(&esqpanel_external_panel_server::on_open, this, _1),
-					std::bind(&esqpanel_external_panel_server::on_message, this, _1, _2, _3),
-					std::bind(&esqpanel_external_panel_server::on_close, this, _1, _2, _3),
-					std::bind(&esqpanel_external_panel_server::on_error, this, _1, _2)
-					);
-		}
-	}
 
-	virtual ~esqpanel_external_panel_server()
+	class external_panel
 	{
-		if (m_server->is_active()) {
-			m_server->remove_endpoint("/esqpanel/socket");
-		}
-	}
-
-	void send_to_all(char c)
-	{
-		// printf("server: send_to_all(%02x)\n", ((unsigned int) c) & 0xff);
-		std::lock_guard<std::recursive_mutex> lock(m_mutex);
-		// printf("server: sending '%02x' to all\n", ((unsigned int) c) & 0xff);
-		m_to_send.str("");
-		m_to_send.put('D');
-		m_to_send.put(c);
-		const std::string &s = m_to_send.str();
-
-		for (auto iter: m_panels)
+	public:
+		static int get_message_type(const char c)
 		{
-			external_panel_ptr panel = iter.second;
-			if (panel->send_display_data())
+			switch(c)
 			{
-				send(iter.first, s);
+			case 'A':
+				return message_type::ANALOG;
+			case 'B':
+				return message_type::BUTTON;
+			case 'C':
+				return message_type::CONTROL;
+			case 'D':
+				return message_type::DISPLAY;
+			case 'I':
+				return message_type::INFO;
+			default:
+				return message_type::UNKNOWN;
 			}
 		}
-	}
 
-	void on_open(http_manager::websocket_connection_ptr connection)
+		external_panel() : m_send_message_types(0)
+		{
+			// printf("session: constructed\n");
+		}
+
+		int handle_control_message(const std::string &command)
+		{
+			int old = m_send_message_types;
+			std::istringstream is(command);
+			if (get_message_type(is.get()) != message_type::CONTROL)
+			{
+				return 0;
+			}
+
+			int n;
+			while (!is.eof()) {
+				char c = is.get();
+				int message_type = external_panel::get_message_type(c);
+				is >> n;
+				int send = (n != 0);
+				if (send)
+				{
+					m_send_message_types |= message_type;
+				}
+				else
+				{
+					m_send_message_types &= ~message_type;
+				}
+			}
+
+			return m_send_message_types ^ old;
+		}
+
+		int send_message_types()
+		{
+			return m_send_message_types;
+		}
+
+		bool send_display_data()
+		{
+			return m_send_message_types & message_type::DISPLAY;
+		}
+
+		bool send_analog_values()
+		{
+			return m_send_message_types & message_type::ANALOG;
+		}
+
+		bool send_buttons()
+		{
+			return m_send_message_types & message_type::BUTTON;
+		}
+
+	private:
+		int m_send_message_types;
+	};
+
+	class external_panel_server
 	{
-		using namespace std::placeholders;
-
-		std::lock_guard<std::recursive_mutex> lock(m_mutex);
-		m_panels[connection] = std::make_shared<external_panel>();
-	}
-
-	void on_message(http_manager::websocket_connection_ptr connection, const std::string &payload, int opcode)
-	{
-		external_panel_ptr panel = external_panel_for_connection(connection);
-		const std::string &command = payload;
-
-		int t = external_panel::get_message_type(command.front());
-
-		if (t == message_type::CONTROL)
+	public:
+		enum websocket_opcode {
+			text = 1,
+			binary = 2
+		};
+		external_panel_server(http_manager *webserver) :
+			m_server(webserver),
+			m_keyboard("unknown"),
+			m_version("1")
 		{
-			int changed = panel->handle_control_message(command);
-			// printf("server: control message, changed = '%x'\n", changed);
-			if ((changed & message_type::DISPLAY) && panel->send_display_data())
-			{
-				// printf("server: control message, sending contents\n");
-				send_contents(connection);
-			}
-
-			if ((changed & message_type::ANALOG) && panel->send_analog_values())
-			{
-				// printf("server: control message, sending analog values\n");
-				send_analog_values(connection);
-			}
-
-			if ((changed & message_type::BUTTON) && panel->send_buttons())
-			{
-				// printf("server: control message, sending button states\n");
-				send_button_states(connection);
+			using namespace std::placeholders;
+			if (m_server->is_active()) {
+				m_server->add_endpoint("/esqpanel/socket",
+						std::bind(&external_panel_server::on_open, this, _1),
+						std::bind(&external_panel_server::on_message, this, _1, _2, _3),
+						std::bind(&external_panel_server::on_close, this, _1, _2, _3),
+						std::bind(&external_panel_server::on_error, this, _1, _2)
+						);
 			}
 		}
-		else if (t == message_type::INFO)
-		{
-			std::ostringstream o;
-			o << "I" << get_keyboard() << "," << get_version();
-			send(connection, o.str());
-		}
-		else
-		{
-			{
-				std::lock_guard<std::recursive_mutex> lock(m_mutex);
-				m_commands.emplace_back(command);
-			}
 
-			// Echo the non-command message to any other connected panels that want it
+		virtual ~external_panel_server()
+		{
+			if (m_server->is_active()) {
+				m_server->remove_endpoint("/esqpanel/socket");
+			}
+		}
+
+		void send_to_all(char c)
+		{
+			// printf("server: send_to_all(%02x)\n", ((unsigned int) c) & 0xff);
+			std::lock_guard<std::recursive_mutex> lock(m_mutex);
+			// printf("server: sending '%02x' to all\n", ((unsigned int) c) & 0xff);
+			m_to_send.str("");
+			m_to_send.put('D');
+			m_to_send.put(c);
+			const std::string &s = m_to_send.str();
+
 			for (auto iter: m_panels)
 			{
-				external_panel_ptr other_panel = iter.second;
-				if (other_panel != panel && (t & other_panel->send_message_types()) != 0)
+				external_panel_ptr panel = iter.second;
+				if (panel->send_display_data())
 				{
-					send(iter.first, command);
+					send(iter.first, s);
 				}
 			}
 		}
-	}
 
-	void on_close(http_manager::websocket_connection_ptr connection, int status, const std::string& reason)
-	{
-		std::lock_guard<std::recursive_mutex> lock(m_mutex);
-		m_panels.erase(connection);
-	}
-
-	void on_error(http_manager::websocket_connection_ptr connection, const std::error_code& error_code)
-	{
-		std::lock_guard<std::recursive_mutex> lock(m_mutex);
-		m_panels.erase(connection);
-	}
-
-	void on_document_request(http_manager::http_request_ptr request, http_manager::http_response_ptr response, const std::string &filename)
-	{
-		m_server->serve_document(request, response, filename);
-	}
-
-	void on_template_request(http_manager::http_request_ptr request, http_manager::http_response_ptr response, const std::string &filename)
-	{
-		using namespace std::placeholders;
-		m_server->serve_template(request, response, filename, std::bind(&esqpanel_external_panel_server::get_template_value, this, _1), '$', '$');
-	}
-
-	external_panel_ptr external_panel_for_connection(http_manager::websocket_connection_ptr connection)
-	{
-		auto it = m_panels.find(connection);
-
-		if (it == m_panels.end()) {
-			// this connection is not in the list. This really shouldn't happen
-			// and probably means something else is wrong.
-			throw std::invalid_argument("No panel avaliable for connection");
-		}
-
-		return it->second;
-	}
-
-	bool has_commands()
-	{
-		// printf("server: has_commands()\n");
-		std::lock_guard<std::recursive_mutex> lock(m_mutex);
-		return !m_commands.empty();
-	}
-
-	std::string get_next_command()
-	{
-		// printf("server: get_next_command()\n");
-		std::lock_guard<std::recursive_mutex> lock(m_mutex);
-		std::string command = std::move(m_commands.front());
-		m_commands.pop_front();
-		return command;
-	}
-
-	void set_index(const std::string &index)
-	{
-		m_index = index;
-	}
-
-	void add_http_document(const std::string &path, const std::string &filename)
-	{
-		m_server->remove_http_handler(path);
-		if (filename != "")
+		void on_open(http_manager::websocket_connection_ptr connection)
 		{
 			using namespace std::placeholders;
-			m_server->add_http_handler(path, std::bind(&esqpanel_external_panel_server::on_document_request, this, _1, _2, filename));
+
+			std::lock_guard<std::recursive_mutex> lock(m_mutex);
+			m_panels[connection] = std::make_shared<external_panel>();
 		}
-	}
 
-	void add_http_template(const std::string &path, const std::string &filename)
-	{
-		m_server->remove_http_handler(path);
-		if (filename != "")
+		void on_message(http_manager::websocket_connection_ptr connection, const std::string &payload, int opcode)
 		{
-			using namespace std::placeholders;
-			m_server->add_http_handler(path, std::bind(&esqpanel_external_panel_server::on_template_request, this, _1, _2, filename));
-		}
-	}
+			external_panel_ptr panel = external_panel_for_connection(connection);
+			const std::string &command = payload;
 
-	void set_content_provider(std::function<bool(std::ostream&)> provider)
-	{
-		m_content_provider = provider;
-	}
+			int t = external_panel::get_message_type(command.front());
 
-	void set_keyboard(const std::string &keyboard)
-	{
-		m_keyboard = keyboard;
-	}
-
-	const std::string &get_keyboard() const
-	{
-		return m_keyboard;
-	}
-
-	const std::string &get_version() const
-	{
-		return m_version;
-	}
-
-	bool get_template_value(std::string &s)
-	{
-		if (s == "keyboard")
-		{
-			s = m_keyboard;
-			return true;
-		}
-		else if (s == "version")
-		{
-			s = m_version;
-			return true;
-		}
-		else
-		{
-			return false;
-		}
-	}
-
-private:
-	void send(http_manager::websocket_connection_ptr connection, const std::string &s)
-	{
-		connection->send_message(s, websocket_opcode::binary);
-	}
-
-	void send_contents(http_manager::websocket_connection_ptr connection)
-	{
-		if (m_content_provider)
-		{
-			m_to_send.str("");
-			m_to_send.put('D');
-			if (m_content_provider(m_to_send))
+			if (t == message_type::CONTROL)
 			{
-				send(connection, m_to_send.str());
+				int changed = panel->handle_control_message(command);
+				// printf("server: control message, changed = '%x'\n", changed);
+				if ((changed & message_type::DISPLAY) && panel->send_display_data())
+				{
+					// printf("server: control message, sending contents\n");
+					send_contents(connection);
+				}
+
+				if ((changed & message_type::ANALOG) && panel->send_analog_values())
+				{
+					// printf("server: control message, sending analog values\n");
+					send_analog_values(connection);
+				}
+
+				if ((changed & message_type::BUTTON) && panel->send_buttons())
+				{
+					// printf("server: control message, sending button states\n");
+					send_button_states(connection);
+				}
+			}
+			else if (t == message_type::INFO)
+			{
+				std::ostringstream o;
+				o << "I" << get_keyboard() << "," << get_version();
+				send(connection, o.str());
+			}
+			else
+			{
+				{
+					std::lock_guard<std::recursive_mutex> lock(m_mutex);
+					m_commands.emplace_back(command);
+				}
+
+				// Echo the non-command message to any other connected panels that want it
+				for (auto iter: m_panels)
+				{
+					external_panel_ptr other_panel = iter.second;
+					if (other_panel != panel && (t & other_panel->send_message_types()) != 0)
+					{
+						send(iter.first, command);
+					}
+				}
 			}
 		}
-	}
 
-	void send_analog_values(http_manager::websocket_connection_ptr connection)
-	{
-		// TODO(cbrunschen): get the current analog values and send them
-	}
+		void on_close(http_manager::websocket_connection_ptr connection, int status, const std::string& reason)
+		{
+			std::lock_guard<std::recursive_mutex> lock(m_mutex);
+			m_panels.erase(connection);
+		}
 
-	void send_button_states(http_manager::websocket_connection_ptr connection)
-	{
-		// TODO(cbrunschen): track current button states and send them
-	}
+		void on_error(http_manager::websocket_connection_ptr connection, const std::error_code& error_code)
+		{
+			std::lock_guard<std::recursive_mutex> lock(m_mutex);
+			m_panels.erase(connection);
+		}
 
-	http_manager *m_server;
-	std::recursive_mutex m_mutex;
+		void on_document_request(http_manager::http_request_ptr request, http_manager::http_response_ptr response, const std::string &filename)
+		{
+			m_server->serve_document(request, response, filename);
+		}
 
-	connection_to_panel_map m_panels;
-	std::list<std::string> m_commands;
-	std::thread m_working_thread;
-	std::ostringstream m_to_send;
+		void on_template_request(http_manager::http_request_ptr request, http_manager::http_response_ptr response, const std::string &filename)
+		{
+			using namespace std::placeholders;
+			m_server->serve_template(request, response, filename, std::bind(&external_panel_server::get_template_value, this, _1), '$', '$');
+		}
 
-	std::string m_index;
-	std::string m_keyboard;
-	std::string m_version;
-	std::function<bool(std::ostream&)> m_content_provider;
-	std::map<const std::string, const std::string> m_template_values;
-};
+		external_panel_ptr external_panel_for_connection(http_manager::websocket_connection_ptr connection)
+		{
+			auto it = m_panels.find(connection);
+
+			if (it == m_panels.end()) {
+				// this connection is not in the list. This really shouldn't happen
+				// and probably means something else is wrong.
+				throw std::invalid_argument("No panel avaliable for connection");
+			}
+
+			return it->second;
+		}
+
+		bool has_commands()
+		{
+			// printf("server: has_commands()\n");
+			std::lock_guard<std::recursive_mutex> lock(m_mutex);
+			return !m_commands.empty();
+		}
+
+		std::string get_next_command()
+		{
+			// printf("server: get_next_command()\n");
+			std::lock_guard<std::recursive_mutex> lock(m_mutex);
+			std::string command = std::move(m_commands.front());
+			m_commands.pop_front();
+			return command;
+		}
+
+		void set_index(const std::string &index)
+		{
+			m_index = index;
+		}
+
+		void add_http_document(const std::string &path, const std::string &filename)
+		{
+			m_server->remove_http_handler(path);
+			if (filename != "")
+			{
+				using namespace std::placeholders;
+				m_server->add_http_handler(path, std::bind(&external_panel_server::on_document_request, this, _1, _2, filename));
+			}
+		}
+
+		void add_http_template(const std::string &path, const std::string &filename)
+		{
+			m_server->remove_http_handler(path);
+			if (filename != "")
+			{
+				using namespace std::placeholders;
+				m_server->add_http_handler(path, std::bind(&external_panel_server::on_template_request, this, _1, _2, filename));
+			}
+		}
+
+		void set_content_provider(std::function<bool(std::ostream&)> provider)
+		{
+			m_content_provider = provider;
+		}
+
+		void set_keyboard(const std::string &keyboard)
+		{
+			m_keyboard = keyboard;
+		}
+
+		const std::string &get_keyboard() const
+		{
+			return m_keyboard;
+		}
+
+		const std::string &get_version() const
+		{
+			return m_version;
+		}
+
+		bool get_template_value(std::string &s)
+		{
+			if (s == "keyboard")
+			{
+				s = m_keyboard;
+				return true;
+			}
+			else if (s == "version")
+			{
+				s = m_version;
+				return true;
+			}
+			else
+			{
+				return false;
+			}
+		}
+
+	private:
+		void send(http_manager::websocket_connection_ptr connection, const std::string &s)
+		{
+			connection->send_message(s, websocket_opcode::binary);
+		}
+
+		void send_contents(http_manager::websocket_connection_ptr connection)
+		{
+			if (m_content_provider)
+			{
+				m_to_send.str("");
+				m_to_send.put('D');
+				if (m_content_provider(m_to_send))
+				{
+					send(connection, m_to_send.str());
+				}
+			}
+		}
+
+		void send_analog_values(http_manager::websocket_connection_ptr connection)
+		{
+			// TODO(cbrunschen): get the current analog values and send them
+		}
+
+		void send_button_states(http_manager::websocket_connection_ptr connection)
+		{
+			// TODO(cbrunschen): track current button states and send them
+		}
+
+		http_manager *m_server;
+		std::recursive_mutex m_mutex;
+
+		connection_to_panel_map m_panels;
+		std::list<std::string> m_commands;
+		std::thread m_working_thread;
+		std::ostringstream m_to_send;
+
+		std::string m_index;
+		std::string m_keyboard;
+		std::string m_version;
+		std::function<bool(std::ostream&)> m_content_provider;
+		std::map<const std::string, const std::string> m_template_values;
+	};
+	
+}  // namespace esqpanel
 
 //**************************************************************************
 //  MACROS / CONSTANTS
@@ -423,7 +427,7 @@ void esqpanel_device::device_start()
 	m_write_tx.resolve_safe();
 	m_write_analog.resolve_safe();
 
-	m_external_panel_server = new esqpanel_external_panel_server(machine().manager().http());
+	m_external_panel_server = new esqpanel::external_panel_server(machine().manager().http());
 	if (machine().manager().http()->is_active()) {
 		m_external_panel_server->set_keyboard(owner()->shortname());
 		m_external_panel_server->set_index("/esqpanel/FrontPanel.html");
